single read only transaction for ticket category availability
